+++ .gitignore
@@ -20,5 +20,5 @@ src/main/java/backend/Task.class
 src/main/java/backend/ToDoList.class
 src/main/java/frontend/Format.class
 data/sirDuke.txt
-src/main/java/backend/Task/Task.class
+src/main/java/backend/task/Task.class
 text-ui-test/level1input.txt
