+++ src/main/java/backend/Task/EventTask.java
@@ -1,4 +1,6 @@
-package backend.Task;
+package backend.task;
+import backend.IllegalStartAndEndDateException;
+
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
 import java.time.format.DateTimeParseException;
